Create generatePrintview()
Create a function to turn the colors table with dropdowns into a new, separate table filled in with text for the colors instead of dropdown menus. Also applies the printview.css to the document in a weird way, but I give up on trying to make the controller do it without refreshing the page.